support local cache by caffeine
