+++ cacheable-caffeine/src/test/resources/application.conf
@@ -1,0 +1,4 @@
+caffeine {
+  maximumSize = 100
+  expireAfterWriteSeconds = 60
+}
