--- cacheable/src/main/scala/org/bitlap/cacheable/ZStreamUpdateCache.scala
@@ -1,40 +1,0 @@
-/*
- * Copyright (c) 2022 bitlap
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
- * the Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
- * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package org.bitlap.cacheable
-
-import zio.stream.ZStream
-
-/**
- * Redis Update Cache for ZStream.
- *
- * @author 梦境迷离
- * @version 2.0,2022/3/19
- */
-trait ZStreamUpdateCache[R, E, T] extends Cache[ZStream[R, E, T]] {
-
-  override final def getIfPresent(business: => ZStream[R, E, T])(identities: List[String], args: List[_]): ZStream[R, E, T] = throw new UnsupportedOperationException()
-
-  override def evict(business: => ZStream[R, E, T])(identities: List[String]): ZStream[R, E, T]
-
-  override def toString: String = "ZStreamUpdateCache"
-
-}
